requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
distlib==0.3.1
Faker==5.6.5
filelock==3.0.12
identify==1.5.13
idna==2.10
iniconfig==1.1.1
nodeenv==1.5.0
packaging==20.8
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-html==3.1.1
pytest-testrail==2.9.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.0
webdriver-manager==3.2.2
